refactor(analytics): extract ChampionWinRateComputeService from the god-file (P10.1)
Pulls win-rate + tier-list computation (raw and stats-backed paths) out of
the 1217-line ChampionAnalyticsComputeService into a focused
ChampionWinRateComputeService (IChampionWinRateComputeService): ComputeWinRates,
ComputeTierList, their *FromStats fast paths, GetAnalyticsComputedAt, plus the
win-rate-only helpers (BuildScopedMatchIdQuery, ApplyStatScope, LookupBanRate,
HasStats).

The adaptive sample-size rule (GetAdaptiveMinimumGamesRequired) was shared with
the builds path, so it moves to a small static AnalyticsSampleThreshold both
services call — no duplication that could let the two surfaces disagree on what
counts as enough data.

ChampionAnalyticsService routes win-rate/tier-list/freshness to the new
dependency; builds/pro/matchups stay on ChampionAnalyticsComputeService.
PrecomputedAnalyticsRefresher is untouched. Pure, behavior-preserving move —
gated by the analytics suite (raw correctness + raw-vs-stats equivalence + the
P7.1 pick/ban assertions). No contract change.

Build clean (no new warnings); Service.Core 183/183, WebAPI 40/40 green.
